Consolidate field test scripts
