+++ .changeset/five-jars-pretend.md
@@ -1,0 +1,9 @@
+---
+'@keystonejs/api-tests': patch
+'@keystonejs/fields': patch
+'@keystonejs/fields-color': patch
+'@keystonejs/fields-markdown': patch
+'@keystonejs/fields-wysiwyg-tinymce': patch
+---
+
+Updated test scripts to be more consistent.
