+++ packages/fields-markdown/src/test-fixtures.js
@@ -32,7 +32,7 @@ export const filterTests = withKeystone => {
     expect(
       await getItems({
         keystone,
-        listKey: 'test',
+        listKey: 'Test',
         where,
         returnFields: 'name content',
         sortBy,
