+++ packages/fields-wysiwyg-tinymce/src/test-fixtures.js
@@ -48,7 +48,7 @@ export const filterTests = withKeystone => {
     expect(
       await getItems({
         keystone,
-        listKey: 'test',
+        listKey: 'Test',
         where,
         returnFields: 'name content',
         sortBy,
