+++ packages/fields/src/types/CalendarDay/test-fixtures.js
@@ -31,7 +31,7 @@ export const filterTests = withKeystone => {
     expect(
       await getItems({
         keystone,
-        listKey: 'test',
+        listKey: 'Test',
         where,
         returnFields: 'name birthday',
         sortBy: 'name_ASC',
