+++ packages/fields/src/types/Checkbox/test-fixtures.js
@@ -30,7 +30,7 @@ export const filterTests = withKeystone => {
     expect(
       await getItems({
         keystone,
-        listKey: 'test',
+        listKey: 'Test',
         where,
         returnFields: 'name enabled',
         sortBy: 'name_ASC',
