+++ packages/fields/src/types/DateTime/test-fixtures.js
@@ -31,7 +31,7 @@ export const filterTests = withKeystone => {
     expect(
       await getItems({
         keystone,
-        listKey: 'test',
+        listKey: 'Test',
         where,
         returnFields: 'name lastOnline',
         sortBy,
