+++ packages/fields/src/types/DateTimeUtc/test-fixtures.js
@@ -31,7 +31,7 @@ export const filterTests = withKeystone => {
     expect(
       await getItems({
         keystone,
-        listKey: 'test',
+        listKey: 'Test',
         where,
         returnFields: 'name lastOnline',
         sortBy,
