+++ packages/fields/src/types/Float/test-fixtures.js
@@ -35,7 +35,7 @@ export const filterTests = withKeystone => {
     expect(
       await getItems({
         keystone,
-        listKey: 'test',
+        listKey: 'Test',
         where,
         returnFields: 'name stars',
         sortBy: 'name_ASC',
