+++ packages/fields/src/types/Select/test-fixtures.js
@@ -65,7 +65,7 @@ export const filterTests = withKeystone => {
     expect(
       await getItems({
         keystone,
-        listKey: 'test',
+        listKey: 'Test',
         where,
         returnFields,
         sortBy: 'name_ASC',
