+++ packages/fields/src/types/Text/test-fixtures.js
@@ -37,7 +37,7 @@ export const filterTests = withKeystone => {
     expect(
       await getItems({
         keystone,
-        listKey: 'test',
+        listKey: 'Test',
         where,
         returnFields: 'name text',
         sortBy,
