+++ packages/fields/tests/test-fixtures.js
@@ -31,7 +31,7 @@ export const filterTests = withKeystone => {
     expect(
       await getItems({
         keystone,
-        listKey: 'test',
+        listKey: 'Test',
         where,
         returnFields: 'id name',
         sortBy: 'name_ASC',
